option to apply model in patches
This option can be used to denoise massive images, even when a single
forward pass of the entire image would use too much memory (presumably
because of all the extra intermediate channels).

This implementation is currently not perfect, since the deep model has
an infinite receptive field. This is something I plan to fix by
switching to batch normalization in the future.